+++ src/containers/Feed/components/SearchInput.tsx
@@ -1,0 +1,19 @@
+import React, { InputHTMLAttributes, ReactNode } from 'react'
+
+interface Props extends InputHTMLAttributes<HTMLInputElement> {}
+
+const SearchInput: React.FC<Props> = ({ ...props }) => {
+  return (
+    <div>
+      <div className="p-1 mb-3 dark:text-white">🔎 Search</div>
+      <input
+        className="rounded-2xl px-5 py-2 mb-8 w-full bg-gray-200 dark:bg-zinc-700 dark:text-white focus:bg-white focus:shadow-md outline-none transition"
+        type="text"
+        placeholder="search keyword.."
+        {...props}
+      />
+    </div>
+  )
+}
+
+export default SearchInput
